calcCbw

Calculates the waterline block coefficient (Cbw) (dimensionless) using actual draft.

calcCbw(Cb, actualDraft, maxDraft)

Arguments

Cb
Maximum block coefficient (vector of numericals, dimensionless) (see calcCb)

actualDraft
Actual draft (vector of numericals, m)

maxDraft
Maximum summer load line draft (vector of numericals, m)

Value

Cbw (vector of numericals, dimensionless)

Details

The block coefficient is the ratio of the vessel's displacement and its volume, defined by its waterline length (lwl), breadth, and draft: $$Cb=\frac{displacement}{lwl*breadth*draft}$$

This function returns the waterline block coefficient (Cbw), which represents a loaded condition estimation using actual draft using the Riddlesworth method. Actual draft is typically obtained from sources such as AIS messages or ship records.

NOTE: Technically, no block coefficient can be greater than 1. Even a waterline block coefficient of 1 itself is extremely improbable and can cause numerical errors elsewhere in the model. Therefore, the upper bound of the waterline block coefficient is fixed at 0.99 in this function.

References

MAN Energy Solutions. 2011. "Basic Principles of Propulsion."

See also

Examples

calcCbw(c(0.82,0.66),c(12.48,11.09),c(13.57,11.49))

#> [1] 0.8149052 0.6559604

calcCbw(0.82,12.48,13.57)

#> [1] 0.8149052