calcHMThrustFactor

Calculate thrust deduction factor (thrustFactor) (dimensionless) from the Holtrop & Mennen method.

calcHMThrustFactor(
  breadth,
  lwl,
  maxDraft,
  maxDisplacement,
  nProp,
  Cp,
  propDiam,
  Cbw,
  Cstern = 0,
  lcb = 0,
  seawaterDensity = 1.025
)

Arguments

breadth
Moulded breadth (vector of numericals, m)

lwl
Waterline length (vector of numericals, m) (see calclwl)

maxDraft
Maximum summer load line draft (vector of numericals, m)

maxDisplacement
Maximum ship displacement (vector of numericals, m^3)

nProp
Number of propellers (vector of numericals, see calcPropNum)

Cp
Prismatic coefficient (vector of numericals, dimensionless) (see calcCp)

propDiam
Propeller diameter (vector of numericals, m) (see calcPropDia)

Cbw
Waterline block coefficient (vector of numericals, dimensionless) (see calcCbw)

Cstern
Afterbody form coefficient:

U-Shaped Hull = 10

Normal Hull = 0 (default)

Can supply either a vector of numericals, a single number, or rely on the default

lcb
Longitudinal position of center of buoyancy (vector of numericals, see calclcb)

seawaterDensity
Sea water density. Default = 1.025 (g/cm^3). Can supply either a vector of numericals, a single number, or rely on the default

Value

thrustFactor (vector of numericals, dimensionless)

Details

Thrust deduction factor is a component of hull efficiency as well as a component of propeller efficiency. It describes the increase in resistance on the hull from water getting sucked back towards the propeller.

Note: In "A Statistical Re-Analysis of Resistance and Propulsion Data", the authors re-analyze with the inclusion of Series 64 hull forms for a total of 334 models included in the analysis. They suggest an update of the single screw thrust equation but that the original equations should be used for twin screw ships.

Additionally, we are assuming conventional stern for all single screw ships. Holtrop & Mennen also include an estimation for thrust factor for single screw ships with open stern for fast sailing ships, but that is not included here.

References

Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance and Propulsion Data'.

See also

Examples

calcHMThrustFactor(c(32.25,32.20),c(218.75,209.25),c(13.57,11.49),c(80097,52382.04),
                   c(1,1),c(0.81,0.67),c(6.7,7),c(0.81,0.65))

#> [1] 0.01698051 0.01861834

calcHMThrustFactor(32.25,218.75,13.57,80097,
                   1,0.81,6.7,0.81)

#> [1] 0.01698051