calcHMTotalRes

Calculate total ship resistance (Rtot) (kN) using the Holtrop & Mennen method.

calcHMTotalRes(
  Rapp,
  Rw,
  Rb,
  Rtr,
  seawaterDensity,
  wettedSA,
  shipSpeed,
  Cf,
  formFactor,
  Ca,
  serviceMargin
)

Arguments

Rapp
Appendage resistance (vector of numericals, kN) (see calcHMAppendageRes)

Rw
Wave waking (similar to residual) resistance (vector of numericals, kN) (see calcHMWaveMakingRes)

Rb
Bulbous bow resistance (vector of numericals, kN) (see calcHMBulbousBowRes)

Rtr
Immersed transom resistance (vector of numericals, kN) (see calcHMImmersedTransomRes)

seawaterDensity
Sea water density. Default = 1.025 (g/cm^2). Can supply either a vector of numericals, a single number, or rely on the default

wettedSA
Wetted hull surface area (vector of numericals, m^2) (see calcHMWettedSA)

shipSpeed
Ship actual speed (vector of numericals, m/s) (see calcSpeedUnitConversion)

Cf
Frictional resistance coefficient (vector of numericals, dimensionless) (see calcCf)

formFactor
Form factor (1+k) (vector of numericals, dimensionless) (see calcHMFormFactor)

Ca
Incremental hull (roughness) resistance coefficient (vector of numericals, dimensionless) (see calcHMCa)

serviceMargin
A service margin to account for weather and sea effects:

At-sea operations = 15 (default)

. Can supply either a vector of numericals, a single number, or rely on the default

Value

Rtot (vector of numericals, kN)

Details

Note that service margin is included here as a resistance term.

References

Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction method." International Shipbuilding Progress 29.

See also

Examples

calcHMTotalRes(Rapp=0.24,
               Rw=2.6,
               Rb=c(4.538188e-07,3.579217e-06,1.180116e-05,2.709520e-05,5.085940e-05),
               Rtr=c(0.58,2.2,4.8,8.1,12.1),
               seawaterDensity=1.025,
               wettedSA=10746.282,
               shipSpeed=seq(1,5,1),
               Cf=c(0.0019,0.0017,0.0016,0.0016,0.0015),
               formFactor=1.275601,
               Ca=0.0003356088,
               serviceMargin=15)

#> [1]  21.40896  69.23655 144.25602 253.41659 373.28878