calcKristPwr

Calculate instantaneous main engine power (kW) using the Kristensen 2016 updates of Harvald equations.

calcKristPwr(
  totalInstalledPwr,
  shipSpeed,
  actualDraft,
  maxDraft,
  shipType,
  lwl,
  breadth,
  maxDisplacement,
  Cb,
  nProp,
  dwt,
  serviceMargin = 15,
  shaftEff = 0.98,
  relRotationEff = 1,
  seawaterTemp = 15,
  seawaterDensity = 1.025,
  pwrUpperBoundPercent = 1,
  pwrLowerBoundPercent = 0.02,
  CmEquationType = "kristensen",
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier"),
  tugShipTypes = c("service.tug", "tug"),
  roroPaxShipTypes = c("passenger", "ferry.pax", "ferry.ro.pax", "cruise", "cruise.ed",
    "yacht", "ro.ro"),
  gCargoShipTypes = c("general.cargo"),
  containerShipTypes = c("container.ship")
)

Arguments

totalInstalledPwr
Total installed main engine power (vector of numericals, kW) (maximum continuous rated power)

shipSpeed
Ship actual speed (vector of numericals, m/s) (see calcSpeedUnitConversion)

actualDraft
Actual draft (vector of numericals, m)

maxDraft
Maximum summer load line draft (vector of numericals, m)

shipType
Ship type (vector of strings, see calcShipType), determined by Stat 5 code:

"container.ship"

"bulk.carrier"

"tanker"

"general.cargo"

"vehicle.carrier"

"reefer"

"ro.ro"

"passenger"

"tug"

"misc"

lwl
Waterline length (vector of numericals, m) (see calclwl)

breadth
Moulded breadth (vector of numericals, m)

maxDisplacement
Maximum ship displacement (vector of numericals, m^3)

Cb
Maximum block coefficient (vector of numericals, dimensionless) (see calcCb)

nProp
Number of propellers (vector of numericals, see calcPropNum)

dwt
Ship maximum deadweight tonnage (vector of numericals, tonnage)

serviceMargin
A service margin to account for weather and sea effects:

Coastal operations = 10

Can supply either a vector of numericals, a single number, or rely on the default

shaftEff
Shaft efficiency (dimensionless). Default = 0.98. Ratio of power delivered to the propeller and the brake power delivered by the engine. Can supply either a vector of numericals, a single number, or rely on the default

relRotationEff
Relative rotational efficiency (dimensionless). Default = 1. Accounts for effect of rotational flow of water around propeller. Can supply either a vector of numericals, a single number, or rely on the default

seawaterTemp
Sea water temperature. Default = 15 (degrees Celsius). Can supply either a vector of numericals, a single number, or rely on the default

seawaterDensity
Sea water density. Default = 1.025 (g/cm^3). Can supply either a vector of numericals, a single number, or rely on the default

pwrUpperBoundPercent
Percent of total installed power at which required power is capped. Default = 1, which indicates required power cannot exceed totalInstalledPwr. Can supply either a vector of numericals, a single number, or rely on the default

pwrLowerBoundPercent
Percent of total installed power to act as lower bound for required power. Default = 0.02, which indicates required power cannot go below 2% of totalInstalledPwr. Can supply either a vector of numericals, a single number, or rely on the default

CmEquationType
Type of equation to estimate the midship section coefficient (see calcCm):

"kristensen"

"benford"

"schneekluth"

tankerBulkCarrierShipTypes
Ship types specified in input shipTypes to be modeled as tankers and bulk carriers.

tugShipTypes
Ship types specified in input shipTypes to be modeled as tugs

roroPaxShipTypes
Ship types specified in input shipTypes to be modeled as RORO and passenger ships

gCargoShipTypes
Ship types specified in input shipTypes to be modeled as general cargo

containerShipTypes
Ship types specified in input shipTypes to be modeled as container ships

Value

power (vector of numericals, kW)

Details

Primary method from Kristensen (2013). Estimation of some inputs use methodology from Rakke (2016).

This method this requires ship types to be grouped. Use the tankerBulkCarrierShipTypes, tugShipTypes, roroPaxShipTypes, gCargoShipTypes, containerShipTypes grouping parameters to provide these ship type groupings. Any ship types not included in these groupings will be considered as miscellaneous vessels.

Ship speed and actual draft are typically obtained from sources such as AIS messages or ship records.

References

Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion Power of Ships."

Kristensen, H. O. 2016. "Revision of statistical analysis and determination of regression formulas for main dimensions of container ships based on data from Clarkson."

Kristensen, H. O. "Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo

Rakke, S. G. 2016. "Ship Emissions Calculation from AIS." NTNU.

See also

Examples

calcKristPwr(
totalInstalledPwr=9363,
shipSpeed=10.8,
actualDraft=12.5,
maxDraft=13.6,
shipType="bulk.carrier",
lwl=218,
breadth=32.25,
maxDisplacement=80097,
Cb=0.8162717,
nProp=1,
dwt=70000,
serviceMargin=15,
shaftEff=0.98,
relRotationEff=1,
seawaterTemp=15,
seawaterDensity=1.025,
CmEquationType="kristensen"
)

#> [1] 9363