calcKristTotalRes

Calculate total ship resistance (Rtot) (kN) using the Kristensen method.

calcKristTotalRes(
  wettedSA,
  Cf,
  Cr,
  Ca,
  Caa,
  seawaterDensity,
  shipSpeed,
  serviceMargin
)

Arguments

wettedSA
Wetted hull surface area (vector of numericals, m^2) (see calcKristWettedSA)

Cf
Frictional resistance coefficient (vector of numericals, dimensionless) (see calcCf)

Cr
Residual resistance coefficient (vector of numericals, dimensionless) (see calcKristCr)

Ca
Incremental hull (roughness) resistance coefficient (vector of numericals, dimensionless) (see calcKristCa)

Caa
Air resistance coefficient (vector of numericals, dimensionless) (see calcKristCaa)

seawaterDensity
Sea water density. Default = 1.025 (g/cm^3). Can supply either a vector of numericals, a single number, or rely on the default

shipSpeed
Ship actual speed (vector of numericals, m/s) (see calcSpeedUnitConversion)

serviceMargin
A service margin to account for weather and sea effects:

At-sea operations = 15

Can supply either a vector of numericals, a single number, or rely on the default

Value

Rtot (vector of numericals, kN)

Details

Note that service margin is included here as a resistance term.

References

Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion Power of Ships."

Kristensen, H. O. "Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo

See also

Examples

calcKristTotalRes(10714.62,0.0015,0.00043,6.46e-05,5e-05,1.025,10.8,15)

#> [1] 1505.998