calcKristWakeFrac

Calculate wake fraction (wakeFraction) (dimensionless) using the Kristensen method.

calcKristWakeFrac(
  shipType,
  breadth,
  lwl,
  Cbw,
  propDiam,
  M,
  nProp,
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier")
)

Arguments

shipType
Ship type (vector of strings, see calcShipType)

breadth
Moulded breadth (vector of numericals, m)

lwl
Waterline length (vector of numericals, m) (see calclwl)

Cbw
Waterline block coefficient (vector of numericals, dimensionless) (see calcCbw)

propDiam
Propeller diameter (vector of numericals, m) (see calcPropDia)

M
Fineness/slenderness coefficient (vector of numericals, dimensionless) (see calcShipM)

nProp
Number of propellers (vector of numericals, see calcPropNum)

tankerBulkCarrierShipTypes
Ship types specified in input shipTypes to be modeled as tankers and bulk carriers vessels (vector of strings)

Value

wakeFraction (vector of numericals, dimensionless)

Details

"The speed of advance of the propeller relative to the water in which it is working is lower than the observed speed of the vessel. This difference in speed, expressed as a percentage of the ship speed, is known as the wake fraction coefficient". https://www.wartsila.com/encyclopedia/term/wake-fraction-coefficient

Wake fraction is a component of hull efficiency as well as a component of propeller efficiency.

This method this requires ship types to be grouped. Use the tankerBulkCarrierShipTypes grouping parameters to provide these ship type groupings. Any ship types not included in this grouping will be considered as miscellaneous vessels.

References

Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion Power of Ships."

Kristensen, H. O. "Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo

See also

Examples

calcKristWakeFrac("bulk.carrier",32.25,218.75, 0.8, 6.7,5.2, 1)

#> [1] 0.3199536

calcKristWakeFrac("oil.tanker",32.25,218.75, 0.8, 6.7,5.2, 1)

#> [1] 0.3199536

calcKristWakeFrac(c("bulk.carrier","container.ship"),
                  c(32.25,49),
                  c(218.75,400),
                  c(0.8,0.75),
                  c(6.7,9.6),
                  c(5.2,6.7),
                  c(1,1))

#> [1] 0.3199536 0.2982483