calcResistanceShipPwr

Calculate ship power (kW) for resistance-based power models.

calcResistanceShipPwr(
  Rtot,
  shipSpeed,
  hullEff,
  openWaterEff,
  totalInstalledPwr,
  shaftEff = 0.98,
  relRotationEff = 1,
  pwrUpperBoundPercent = 1,
  pwrLowerBoundPercent = 0.02
)

Arguments

Rtot
Total ship resistance (vector of numericals, kN) (see calcKristTotalRes or calcHMTotalRes)

shipSpeed
Ship actual speed (vector of numericals, m/s) (see calcSpeedUnitConversion)

hullEff
Hull efficiency (vector of numericals, dimensionless) (see calcHullEff)

openWaterEff
Open water efficiency (vector of numericals, dimensionless) (see calcOpenWaterEff)

totalInstalledPwr
Total installed main engine power (vector of numericals, kW) (maximum continuous rated power)

shaftEff
Shaft efficiency (dimensionless). Default = 0.98. Ratio of power delivered to the propeller and the brake power delivered by the engine. Can supply either a vector of numericals, a single number, or rely on the default

relRotationEff
Relative rotational efficiency (dimensionless). Default = 1. Accounts for effect of rotational flow of water around propeller. Can supply either a vector of numericals, a single number, or rely on the default

pwrUpperBoundPercent
Percent of total installed power at which required power is capped (1 indicates required power cannot exceed totalInstalledPwr). Can supply either a vector of numericals, a single number, or rely on the default

pwrLowerBoundPercent
Percent of total installed power to act as lower bound for required power (0.02 indicates required power cannot go below 2% of totalInstalledPwr). Can supply either a vector of numericals, a single number, or rely on the default

Value

power (vector of numericals, kW)

Details

This function is called by both Holtrop & Mennen (calcHMPwr) and Kristensen (calcKristPwr) power calculations, where each uses has its own total resistance calculation (calcHMTotalRes and calcKristTotalRes).

Assumptions for shaft efficiency and relative rotational efficiency are based on MAN.

References

MAN Energy Solutions. 2011. "Basic Principles of Propulsion."

See also

Examples

calcResistanceShipPwr(398.487,10.8, 1.191176, 0.721239,9363, shaftEff=0.98,relRotationEff=1)

#> [1] 5111.598