calcShipType

Assigns a generalizable ship type based on the Vessel_Type field in ship registry data.

calcShipType(vesselType, method = "imo", inputTableLocation = NULL)

Arguments

vesselType
Unstandardized vessel types from ship registry data (vector of strings)

method
Select the methodology that should be used (i.e., which set of ship types to return):

"starcrest"

inputTableLocation
File path (optional). Used to specify a user-supplied mapping between unstandardized VesselTypes and standardized shipTypes. See details for formatting requirements.

Value

shipType, a data.table column of the general ship type, corresponding to the input data. NAs represent Vessel_Types not in shipMap (or the inputTableLocation parameter, if that was used).

Details

There are different sets of ship types available for use: IMO's ship types or Starcrest's ship types. Since ship types are important for assigning default auxiliary and boiler engine assumptions, the methodology selected here should match with the methodology used for assigning auxiliary and boiler loads. The IMO ship types are based on the Third GHG Study, and the Starcrest ship types are based on the 2017 Port of Los Angeles air emissions inventory.

If the default mapping between the unstandardized VesselTypes and standardized shipTypes does not catch everything (i.e, if there are unstandardized VesselTypes that are not standardized by this function), or if you want to use your own standardized shipTypes, you can supply a custom mapping file using the inputTableLocation argument.

If user-supplied ship type mapping is used, the file should be in .csv format with two columns and a header row. The headers should be "Vessel_Type" and "shipType", and the values should specify which Vessel_Types map to which shipTypes. See data(shipMap) for the default mapping used by this function.

For more information about ship type, see Section 3.3.3 of the Port Emissions Inventory Guidance.

References

International Maritime Organization. 2014. "Third IMO GHG study 2014 - Final report." London: International Maritime Organization.

Starcrest Consulting Group. 2018. "Port of Los Angeles Air Emissions Inventory - 2017." APP 171019-517 A.

EPA. 2020. "Port Emissions Inventory Guidance: Methodologies for Estimating Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI: Office of Transportation and Air Quality. US Environmental Protection Agency.

See also

Examples

calcShipType(vesselType = c("Fully Cellular Container",
                            "Deck Cargo Carrier",
                            "Open Hatch Carrier",
                            "Cruise Ship",
                            "LNG Carrier")
             )

#>             imoShipType
#> 1:       container.ship
#> 2:        general.cargo
#> 3:         bulk.carrier
#> 4:               cruise
#> 5: liquified.gas.tanker

calcShipType(vesselType = c("Fully Cellular Container",
                            "Deck Cargo Carrier",
                            "Open Hatch Carrier",
                            "Cruise Ship",
                            "LNG Carrier"),
             method="starcrest")

#>    starcrestShipType
#> 1:    container.ship
#> 2:     general.cargo
#> 3:      bulk.carrier
#> 4:            cruise
#> 5:            tanker